Fix forms sign in sign up (#20)
* change page sign up

* change signup

* add component for elements sign up

* change page sign in

* add component sign in for elements

* change the file in sign up
